% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basic.R
\name{adjustNormalMultinomial}
\alias{adjustNormalMultinomial}
\title{Finds the mean and covariance of a normal multinomial distribution}
\usage{
adjustNormalMultinomial(X)
}
\arguments{
\item{X}{normal-multinomial sample}
}
\description{
Finds the mean and covariance of a normal multinomial distribution
}

