% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjustNormalMultinomial2}
\alias{adjustNormalMultinomial2}
\title{Finds the mean and covariance of a normal multinomial distribution}
\usage{
adjustNormalMultinomial2(X, A, eps, iter, minSigma)
}
\arguments{
\item{X}{normal-multinomial sample}

\item{A}{initial values}
}
\description{
Finds the mean and covariance of a normal multinomial distribution
}

