% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjustNormalMultinomial_internal}
\alias{adjustNormalMultinomial_internal}
\title{Finds the mean and covariance of a normal multinomial distribution}
\usage{
adjustNormalMultinomial_internal(X, A, eps = 1e-04, iter = 100L,
  minSigma = 1e-06)
}
\arguments{
\item{X}{normal-multinomial sample}

\item{A}{initial values}
}
\description{
Finds the mean and covariance of a normal multinomial distribution
}

