% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{dm_expected}
\alias{dm_expected}
\title{Calculate the expected probabilities}
\usage{
dm_expected(X, alpha)
}
\arguments{
\item{X}{count data set}

\item{alpha}{paramater alpha}
}
\value{
the expected probabilities
}
\description{
Calculate the probabilities conditioned on parameters alpha
more likely to generate dataset X
}
\examples{
X = rnormalmultinomial(100, 100, rep(0,4), diag(4))
fit = dm_fit(X)
P = dm_expected(X, fit$alpha)
head(P)
}

