% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{dm_fit}
\alias{dm_fit}
\title{Dirichlet-multinomial parameter estimation.}
\usage{
dm_fit(X, eps = 1e-04, max.iter = 5000, verbose = FALSE)
}
\arguments{
\item{X}{count data set}

\item{eps}{maximum error accepted for alpha parameters}

\item{max.iter}{maximum number of steps allowed}
}
\value{
A list with parameter alpha and the number of iterations before convergence
}
\description{
Dirichlet-multinomial parameter estimation.
}
\examples{
X = rnormalmultinomial(100, 100, rep(0,4), diag(4))
dm_fit(X)
}

