% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrandom.R
\name{dnormalmultinomial}
\alias{dnormalmultinomial}
\title{Probability of a given count}
\usage{
dnormalmultinomial(x, mu, sigma, nsim = 100)
}
\arguments{
\item{x}{counts}

\item{mu}{mean parameter for the logratio normal distribution in ilr coordinates}

\item{sigma}{covariance parameter parameter for the  logratio normal distribution in ilr coordinates}

\item{nsim}{simulations done to calculate the integral}
}
\value{
probability
}
\description{
Probability of a given count
}
\examples{
dnormalmultinomial(c(0,0), mu = c(0,0), sigma = matrix(c(2,1,1,2), nrow=2), nsim = 1000)
}

