% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EM_step_OMP}
\alias{EM_step_OMP}
\title{Compute the mean and covariance of a normal multinomial distribution after itering an EM-algorithm}
\usage{
EM_step_OMP(X, mu, sigma, nsim = 1000L, niter = 1L, nthreads = 1L)
}
\arguments{
\item{X}{normal-multinomial sample}

\item{mu}{initial meean estimate}

\item{sigma}{initial covariance estimate}

\item{nsim}{number of repetitions for the montecarlo integration process}

\item{niter}{number of iteration}
}
\description{
Compute the mean and covariance of a normal multinomial distribution after itering an EM-algorithm
}

