% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{initialize_with_aitchison}
\alias{initialize_with_aitchison}
\title{Approximation to Mu and Sigma parameters obtained
from raw data as proposed by Aitchison.
Expected values are initiated using the obtained Mu}
\usage{
initialize_with_aitchison(X)
}
\description{
Approximation to Mu and Sigma parameters obtained
from raw data as proposed by Aitchison.
Expected values are initiated using the obtained Mu
}

