% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{initialize_with_bootstrap}
\alias{initialize_with_bootstrap}
\title{A bootstrap method is used to obtained estimates of Mu and Sigma
The expected values are initialized using vector Mu}
\usage{
initialize_with_bootstrap(X)
}
\description{
A bootstrap method is used to obtained estimates of Mu and Sigma
The expected values are initialized using vector Mu
}

