% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{initialize_with_bootstrapstep}
\alias{initialize_with_bootstrapstep}
\title{A bootstrap method is used to obtained estimates of Mu and Sigma
The expected values are calculated using one iteration of EM algorithm}
\usage{
initialize_with_bootstrapstep(X, steps = 1, parallel.cluster)
}
\description{
A bootstrap method is used to obtained estimates of Mu and Sigma
The expected values are calculated using one iteration of EM algorithm
}

