% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{initialize_with_dapprox}
\alias{initialize_with_dapprox}
\title{The expected values are obtained from the solution
of an approximation to the parameters of a Dirichlet model (see Dirichlet estimation article)}
\usage{
initialize_with_dapprox(X)
}
\description{
The expected values are obtained from the solution
of an approximation to the parameters of a Dirichlet model (see Dirichlet estimation article)
}

