% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{initialize_with_dm}
\alias{initialize_with_dm}
\title{Initialization using the expected values obtained
from a Dirichlet-Multinomial fitting}
\usage{
initialize_with_dm(X)
}
\description{
Initialization using the expected values obtained
from a Dirichlet-Multinomial fitting
}

