% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLike}
\alias{logLike}
\title{Evaluates the log-likelihood of thenormal multinomial distribution given the hidden observation}
\usage{
logLike(X, A, mu, inv_sigma)
}
\arguments{
\item{X}{normal-multinomial observations}

\item{A}{aln non-observed obeservations}

\item{mu}{mean parameter for the mean in a aln-normal distribution}

\item{sigma}{parameter for the sigma in a aln-normal distribution}
}
\value{
Loglikelihood of the oberserved and non-observed data
}
\description{
Evaluates the log-likelihood of thenormal multinomial distribution given the hidden observation
}

