% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mstep}
\alias{Mstep}
\title{Finds the mean and covariance of a normal multinomial distribution}
\usage{
Mstep(A, mu, inv_sigma, X, eps = 1e-08)
}
\arguments{
\item{A}{aln hidden obeservations}

\item{mu}{mean parameter for the mean in a aln-normal distribution}

\item{X}{normal-multinomial observations}

\item{sigma}{parameter for the sigma in a aln-normal distribution}
}
\value{
Loglikelihood oberserved data
}
\description{
Finds the mean and covariance of a normal multinomial distribution
}

