% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvf_deriv2}
\alias{mvf_deriv2}
\title{Finds the mean and covariance of a normal multinomial distribution}
\usage{
mvf_deriv2(I, J, a, mu, inv_sigma, x)
}
\arguments{
\item{a}{aln hidden obeservation}

\item{mu}{mean parameter for the mean in a aln-normal distribution}

\item{x}{normal-multinomial observation}

\item{sigma}{parameter for the sigma in a aln-normal distribution}
}
\value{
Loglikelihood og oberserved data
}
\description{
Finds the mean and covariance of a normal multinomial distribution
}

