% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{nm_expected}
\alias{nm_expected}
\title{Calculate the expected probabilities P to generate the dataset X conditioned on parameters
mu and sigma}
\usage{
nm_expected(X, mu, sigma, nsim = 1000, parallel.cluster = NULL)
}
\arguments{
\item{X}{count data set}

\item{mu}{paramater mu}

\item{sigma}{parameter sigma}

\item{nsim}{number of simulation used to calculated the expected values}

\item{parallel.cluster}{instance of a Socket Cluster (see function makeCluster from package parallel)}
}
\value{
the expected probabilities
}
\description{
Calculate the expected probabilities P to generate the dataset X conditioned on parameters
mu and sigma
}
\examples{
X = rnormalmultinomial(100, 100, rep(0,4), diag(4))
fit = nm_fit(X, verbose = T)
P = nm_expected(X, fit$mu, fit$sigma)
head(P)
}

