% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalmultinomial_fitting}
\alias{normalmultinomial_fitting}
\title{Finds the mean and covariance of a normal multinomial distribution}
\usage{
normalmultinomial_fitting(X, nsim = 1000L, niter = 20L, prop = 0.66,
  version = 0L, nthreads = 1L)
}
\arguments{
\item{X}{normal-multinomial sample}

\item{nsim}{number of repetitions for the montecarlo integration process}

\item{niter}{number of iterations for the EM-algorithm}

\item{prop}{first 0 imputation}
}
\description{
Finds the mean and covariance of a normal multinomial distribution
}

