% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parlament2015}
\alias{parlament2015}
\title{Results for parliament catalan elections by municipality and municipality variables}
\format{A data frame with 947 rows and 34 variables:
\item{code}{Municipality code}
\item{name}{Munipality name}
\item{jxsi}{Votes to Juns pel Si}
\item{psc}{Votes to Partit Socialista de Catalunyac}
\item{pp}{Votes to Partit Popular}
\item{catsp}{Votes to Catalunya si que es pot}
\item{cs}{Votes to Ciutadans}
\item{cup}{Votes to CUP}
\item{other}{Vote to other parties}
\item{blank}{Blank votes}
\item{null}{Null votes}
\item{abs}{Abstentions}
\item{pop}{Municipality population}
\item{birth.cat}{Number of people born in Catalonia}
\item{birth.esp}{Number of people born in Spain but not in Catalonia}
\item{birth.other}{Number of people born ouside Spain}
\item{ss.agri}{ss.agri}
\item{ss.ind}{ss.ind}
\item{ss.const}{ss.const}
\item{ss.serv}{ss.serv}
\item{ss.aut.agri}{ss.aut.agri}
\item{ss.aut.ind}{ss.aut.ind}
\item{ss.aut.const}{ss.aut.const}
\item{ss.aut.serv}{ss.aut.serv}
\item{pop.m.00_14}{Male population until 14 years old}
\item{pop.m.15_64}{Male population from 15 to 64 years old}
\item{pop.h.65_84}{Male population from 65 to 84 years old}
\item{pop.h.85}{Male population from 85 years old}
\item{pop.w.00_14}{Female population until 14 years old}
\item{pop.w.15_64}{Female population from 15 to 64 years old}
\item{pop.w.65_84}{Female population from 65 to 84 years old}
\item{pop.w.85}{Female population from 85 years old}
\item{lon}{Longitude coordinates}
\item{lat}{Latitude coordinates}}
\source{
\url{http://www.idescat.cat/pub/?id=elepc} \url{http://www.idescat.cat/emex/}
}
\usage{
parlament2015
}
\description{
Votes in each municipality
}
\keyword{datasets}

