% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{rmultinomial}
\alias{rmultinomial}
\title{Simulate variables following a multinomial distribution.}
\usage{
rmultinomial(A, size, seed = 1L)
}
\arguments{
\item{A}{matrix with the probabilities used to generate the sample}

\item{size}{vector with the countings in each row of A}
}
\value{
sample with the same dimension as A
}
\description{
Simulate variables following a multinomial distribution.
}

