% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basic.R
\name{rnorm.multinom}
\alias{rnorm.multinom}
\title{Generate a sample from a normalmultinomial distribution}
\usage{
rnorm.multinom(N, Mu, Sigma, Size)
}
\arguments{
\item{N}{sample size}

\item{Mu}{alr mean indicating the center of the nm distribution}

\item{Sigma}{alt covariance indicating the variable of the nm distribution}

\item{Size}{Sampling size for the counting process}
}
\value{
A sample comming from a normal-multinomial distribution
}
\description{
This function generates a sample generated from a normal-multinomial (nm) distribution
}

