% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rnormal}
\alias{rnormal}
\title{Simulate variables following a normal distribution.}
\usage{
rnormal(n, mu, sigma)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean parameter for the mean in a aln-normal distribution}

\item{sigma}{parameter for the sigma in a aln-normal distribution}
}
\value{
Sample matrix
}
\description{
Simulate variables following a normal distribution.
}

