% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rnormalmultinomial}
\alias{rnormalmultinomial}
\title{Simulate variables following a normal multinomial distribution.}
\usage{
rnormalmultinomial(mu, sigma, size)
}
\arguments{
\item{mu}{mean parameter for the mean in a aln-normal distribution}

\item{sigma}{parameter for the sigma in a aln-normal distribution}

\item{n}{sample size}
}
\value{
Sample matrix
}
\description{
Simulate variables following a normal multinomial distribution.
}

