% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrandom.R
\name{dnormalmultinomial}
\alias{dnormalmultinomial}
\title{Probability of a given count}
\usage{
dnormalmultinomial(x, mu, sigma, nsim = 100)
}
\arguments{
\item{x}{counts}

\item{mu}{mean parameter for the mean in the aln-normal distribution}

\item{sigma}{parameter for the sigma in the aln-normal distribution}

\item{nsim}{simulations done to calculate the integral}
}
\value{
probability
}
\description{
Probability of a given count
}
\examples{
rnormalmultinomial(10, 5, mu = c(0,0), sigma = matrix(c(2,1,1,2), nrow=2))
}

