% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nm_fit}
\alias{nm_fit}
\title{Finds the mean and covariance of a normal multinomial distribution}
\usage{
nm_fit(X, nsim = 100L, niter = 20L, prop = 0.66, version = 0L)
}
\arguments{
\item{X}{normal-multinomial sample}

\item{nsim}{number of repetitions for the montecarlo integration process}

\item{niter}{number of iterations for the EM-algorithm}

\item{prop}{first 0 imputation}
}
\description{
Finds the mean and covariance of a normal multinomial distribution
}

