% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rmultinomial}
\alias{rmultinomial}
\title{Simulate variables following a multinomial distribution.}
\usage{
rmultinomial(n, size, prob)
}
\arguments{
\item{n}{matrix with the probabilities used to generate the sample}

\item{size}{number of trials en each observation}
}
\value{
multinomial sample in a matrix
}
\description{
Simulate variables following a multinomial distribution.
}
\examples{
rmultinomial(10, c(10, 100), c(0.2,0.8))
}

