% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrandom.R
\name{rnormal}
\alias{rnormal}
\title{Simulate variables following a normal distribution.}
\usage{
rnormal(n, mu, sigma)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean parameter for the mean in a normal distribution}

\item{sigma}{parameter for the sigma in a normal distribution}
}
\value{
Sample matrix
}
\description{
Simulate variables following a normal distribution.
}
\examples{
rnormal(10, mu = c(0,0), sigma = matrix(c(2,1,1,2), nrow=2))
}

