% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rnormalmultinomial}
\alias{rnormalmultinomial}
\title{Simulate variables following a normal multinomial distribution.}
\usage{
rnormalmultinomial(mu, sigma, size, seed = 1L)
}
\arguments{
\item{mu}{mean parameter for the mean in a aln-normal distribution}

\item{sigma}{parameter for the sigma in a aln-normal distribution}

\item{size}{number of trials}

\item{n}{number of observations}
}
\value{
Sample matrix
}
\description{
Simulate variables following a normal multinomial distribution.
}

