% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrandom.R
\name{rnormalmultinomial}
\alias{rnormalmultinomial}
\title{Simulate variables following a normal multinomial distribution.}
\usage{
rnormalmultinomial(n, size, mu, sigma, probs = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{size}{number of trials}

\item{mu}{mean parameter for the mean in the aln-normal distribution}

\item{sigma}{parameter for the sigma in the aln-normal distribution}
}
\value{
Sample matrix
}
\description{
Simulate variables following a normal multinomial distribution.
}
\examples{
rnormalmultinomial(10, 5, mu = c(0,0), sigma = matrix(c(2,1,1,2), nrow=2))
}

